combine pie, treemap, circlepack into a single coherent schema.
