--- compiler/src/template-api/json-schema/Pie.json
@@ -1,137 +1,0 @@
-{
-    "type": "object",
-    "$id": "pieSchema",
-    "definitions": {
-        "db": {
-            "type": "string",
-            "minLength": 1
-        },
-        "query": {
-            "type": "object",
-            "properties": {
-                "table": {
-                    "type": "string",
-                    "minLength": 1
-                },
-                "dimensions": {
-                    "type": "array",
-                    "item": {
-                        "type": "string",
-                        "minLength": 1
-                    },
-                    "minItems": 1
-                },
-                "measure": {
-                    "type": "string",
-                    "pattern": "^(AVG|SUM|COUNT|MAX|MIN)\\(.+\\)$"
-                },
-                "sampleFields": {
-                    "type": "array",
-                    "item": {
-                        "type": "string",
-                        "minLength": 1
-                    },
-                    "default": []
-                }
-            },
-            "required": ["table", "dimensions", "measure"],
-            "additionalProperties": false
-        },
-        "tooltip": {
-            "type": "object",
-            "properties": {
-                "columns": {
-                    "type": "array",
-                    "minItems": 1,
-                    "item": {
-                        "type": "string",
-                        "minLength": 1
-                    }
-                },
-                "aliases": {
-                    "type": "array",
-                    "minItems": 1,
-                    "item": {
-                        "type": "string",
-                        "minLength": 1
-                    }
-                }
-            },
-            "required": ["columns", "aliases"],
-            "additionalProperties": false
-        },
-        "colorScheme": {
-            "enum": [
-                "schemeAccent",
-                "schemeCategory10",
-                "schemeDark2",
-                "schemePaired",
-                "schemePastel1",
-                "schemePastel2",
-                "schemeSet1",
-                "schemeSet2",
-                "schemeSet3",
-                "schemeTableau10"
-            ]
-        },
-        "transition": {
-            "type": "boolean"
-        },
-        "legend": {
-            "type": "object",
-            "properties": {
-                "title": {
-                    "type": "string",
-                    "maxLength": 80,
-                    "default": ""
-                },
-                "domain": {
-                    "type": "object",
-                    "default": {}
-                }
-            },
-            "additionalProperties": false
-        },
-        "width": {
-            "type": "number",
-            "minimum": 600,
-            "maximum": 1500
-        },
-        "height": {
-            "type": "number",
-            "minimum": 600,
-            "maximum": 1500,
-            "default": 1000
-        }
-    },
-    "properties": {
-        "db": {"$ref": "#/definitions/db"},
-        "query": {"$ref": "#/definitions/query"},
-        "tooltip": {"$ref": "#/definitions/tooltip"},
-        "radius": {
-            "type": "number",
-            "minimum": 100,
-            "maximum": 400,
-            "default": 300
-        },
-        "colorScheme": {
-            "$ref": "#/definitions/colorScheme",
-            "default": "schemeTableau10"
-        },
-        "transition": {
-            "$ref": "#/definitions/transition",
-            "default": true
-        },
-        "legend": {"$ref": "#/definitions/legend"},
-        "width": {
-            "$ref": "#/definitions/width",
-            "default": 1000
-        },
-        "height": {
-            "$ref": "#/definitions/height",
-            "default": 1000
-        }
-    },
-    "required": ["db", "query"],
-    "additionalProperties": false
-}
