--- compiler/src/template-api/json-schema/StaticHierarchy.json
@@ -1,70 +1,0 @@
-{
-    "type": "object",
-    "definitions": {
-        "padding": {
-            "type": "integer",
-            "minimum": 0,
-            "maximum": 15
-        },
-        "textFields": {
-            "type": "array",
-            "item": {
-                "type": "string",
-                "minLength": 1
-            }
-        },
-        "colorScheme": {
-            "enum": [
-                "interpolateBlues",
-                "interpolateBrBG",
-                "interpolateBuGn",
-                "interpolateBuPu",
-                "interpolateGnBu",
-                "interpolateGreens",
-                "interpolateGreys",
-                "interpolateOrRd",
-                "interpolateOranges",
-                "interpolatePiYG",
-                "interpolatePuBu",
-                "interpolatePuBuGn",
-                "interpolatePuOr",
-                "interpolatePuRd",
-                "interpolatePurples",
-                "interpolateRdBu",
-                "interpolateRdGy",
-                "interpolateRdPu",
-                "interpolateRdYlBu",
-                "interpolateRdYlGn",
-                "interpolateReds",
-                "interpolateSpectral",
-                "interpolateYlGn",
-                "interpolateYlGnBu",
-                "interpolateYlOrBr",
-                "interpolateYlOrRd"
-            ]
-        }
-    },
-    "properties": {
-        "db": {"$ref": "pieSchema#/definitions/db"},
-        "query": {"$ref": "pieSchema#/definitions/query"},
-        "type": {
-            "enum": ["treemap", "circlePack"]
-        },
-        "tooltip": {"$ref": "pieSchema#/definitions/tooltip"},
-        "padding": {"$ref": "#/definitions/padding", "default": 3},
-        "textFields": {"$ref": "#/definitions/textFields", "default": []},
-        "colorScheme": {
-            "$ref": "#/definitions/colorScheme",
-            "default": "interpolateGnBu"
-        },
-        "transition": {
-            "$ref": "pieSchema#/definitions/transition",
-            "default": true
-        },
-        "legend": {"$ref": "pieSchema#/definitions/legend"},
-        "width": {"$ref": "pieSchema#/definitions/width", "default": 1000},
-        "height": {"$ref": "pieSchema#/definitions/height", "default": 1000}
-    },
-    "required": ["db", "query", "type"],
-    "additionalProperties": false
-}
