+++ compiler/src/template-api/json-schema/staticTemplate.json
@@ -1,0 +1,209 @@
+{
+    "type": "object",
+    "properties": {
+        "db": {
+            "type": "string",
+            "minLength": 1
+        },
+        "query": {
+            "type": "object",
+            "properties": {
+                "table": {
+                    "type": "string",
+                    "minLength": 1
+                },
+                "dimensions": {
+                    "type": "array",
+                    "item": {
+                        "type": "string",
+                        "minLength": 1
+                    },
+                    "minItems": 1
+                },
+                "measure": {
+                    "type": "string",
+                    "pattern": "^(AVG|SUM|COUNT|MAX|MIN)\\(.+\\)$"
+                },
+                "sampleFields": {
+                    "type": "array",
+                    "item": {
+                        "type": "string",
+                        "minLength": 1
+                    },
+                    "default": []
+                }
+            },
+            "required": ["table", "dimensions", "measure"],
+            "additionalProperties": false
+        },
+        "type": {
+            "enum": ["pie", "treemap", "circlePack"]
+        },
+        "tooltip": {
+            "type": "object",
+            "properties": {
+                "columns": {
+                    "type": "array",
+                    "minItems": 1,
+                    "item": {
+                        "type": "string",
+                        "minLength": 1
+                    }
+                },
+                "aliases": {
+                    "type": "array",
+                    "minItems": 1,
+                    "item": {
+                        "type": "string",
+                        "minLength": 1
+                    }
+                }
+            },
+            "required": ["columns", "aliases"],
+            "additionalProperties": false
+        },
+        "colorScheme": {},
+        "transition": {
+            "type": "boolean",
+            "default": true
+        },
+        "legend": {
+            "type": "object",
+            "properties": {
+                "title": {
+                    "type": "string",
+                    "maxLength": 80,
+                    "default": ""
+                },
+                "domain": {
+                    "type": "object",
+                    "default": {}
+                }
+            },
+            "additionalProperties": false
+        },
+        "width": {
+            "type": "number",
+            "minimum": 600,
+            "maximum": 1500,
+            "default": 1000
+        },
+        "height": {
+            "type": "number",
+            "minimum": 600,
+            "maximum": 1500,
+            "default": 1000
+        },
+        "textFields": {
+            "type": "array",
+            "item": {
+                "type": "string",
+                "minLength": 1
+            }
+        },
+        "padding": {
+            "type": "integer",
+            "minimum": 0,
+            "maximum": 15
+        },
+        "radius": {
+            "type": "number",
+            "minimum": 100,
+            "maximum": 400
+        }
+    },
+    "required": ["db", "query", "type"],
+    "additionalProperties": false,
+    "allOf": [
+        {
+            "if": {
+                "properties": {
+                    "type": {
+                        "enum": ["pie"]
+                    }
+                }
+            },
+            "then": {
+                "properties": {
+                    "colorScheme": {
+                        "enum": [
+                            "schemeAccent",
+                            "schemeCategory10",
+                            "schemeDark2",
+                            "schemePaired",
+                            "schemePastel1",
+                            "schemePastel2",
+                            "schemeSet1",
+                            "schemeSet2",
+                            "schemeSet3",
+                            "schemeTableau10"
+                        ],
+                        "default": "schemeTableau10"
+                    },
+                    "radius": {
+                        "default": 300
+                    },
+                    "textFields": {
+                        "type": "null"
+                    },
+                    "padding": {
+                        "type": "null"
+                    }
+                }
+            }
+        },
+        {
+            "if": {
+                "properties": {
+                    "type": {
+                        "enum": ["treemap", "circlePack"]
+                    }
+                }
+            },
+            "then": {
+                "properties": {
+                    "colorScheme": {
+                        "enum": [
+                            "interpolateBlues",
+                            "interpolateBrBG",
+                            "interpolateBuGn",
+                            "interpolateBuPu",
+                            "interpolateGnBu",
+                            "interpolateGreens",
+                            "interpolateGreys",
+                            "interpolateOrRd",
+                            "interpolateOranges",
+                            "interpolatePiYG",
+                            "interpolatePuBu",
+                            "interpolatePuBuGn",
+                            "interpolatePuOr",
+                            "interpolatePuRd",
+                            "interpolatePurples",
+                            "interpolateRdBu",
+                            "interpolateRdGy",
+                            "interpolateRdPu",
+                            "interpolateRdYlBu",
+                            "interpolateRdYlGn",
+                            "interpolateReds",
+                            "interpolateSpectral",
+                            "interpolateYlGn",
+                            "interpolateYlGnBu",
+                            "interpolateYlOrBr",
+                            "interpolateYlOrRd"
+                        ],
+                        "default": "interpolateGnBu"
+                    },
+                    "textFields": {
+                        "default": []
+                    },
+                    "padding": {
+                        "default": 3
+                    },
+                    "radius": {
+                        "type": "null"
+                    }
+                }
+            }
+        }
+    ]
+}
